+++ zone/lua_hate_list.h
@@ -2,13 +2,11 @@
 #define EQEMU_LUA_HATE_LIST_H
 #ifdef LUA_EQEMU
 
+#include <sol/sol.hpp>
 #include "lua_ptr.h"
 
 class Lua_Mob;
 struct struct_HateList;
-
-luabind::scope lua_register_hate_entry();
-luabind::scope lua_register_hate_list();
 
 class Lua_HateEntry : public Lua_Ptr<struct_HateList>
 {
@@ -31,6 +29,7 @@ public:
 struct Lua_HateList
 {
 	std::vector<Lua_HateEntry> entries;
+	sol::as_table_t<std::vector<Lua_HateEntry>> get_entries() { return sol::as_table(entries); }
 };
 
 #endif
