+++ zone/lua_inventory.h
@@ -12,12 +12,6 @@ namespace EQ
 	class InventoryProfile;
 }
 
-namespace luabind {
-	struct scope;
-}
-
-luabind::scope lua_register_inventory();
-
 // This class should be deprecated due to the nature of inventory actions.
 // Direct manipulation of the inventory system bypasses the client management
 // of database calls and can lead to lost items, duplicated items and/or
