+++ zone/lua_item.h
@@ -9,12 +9,6 @@ namespace EQ
 	struct ItemData;
 }
 
-namespace luabind {
-	struct scope;
-}
-
-luabind::scope lua_register_item();
-
 class Lua_Item : public Lua_Ptr<const EQ::ItemData>
 {
 	typedef const EQ::ItemData NativeType;
